refactor(utils): move redux extension tools into dedicated dir
