+++ src/utils/redux-extension/index.ts
@@ -1,0 +1,1 @@
+export { useReduxConnector } from './useReduxConnector';
